Really fix a couple of compile warnings on Tru64 5.0

+++ dhcpctl/cltest.c
@@ -40,6 +40,7 @@
 #include <sys/time.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <stdarg.h>
 #include <isc/result.h>
 #include "dhcpctl.h"
@@ -114,7 +115,7 @@ int main (argc, argv)
 	if (name) {
 		status = dhcpctl_new_authenticator (&authenticator,
 						    name, algorithm, pass,
-						    (int)strlen (pass) + 1);
+						    strlen (pass) + 1);
 		if (status != ISC_R_SUCCESS) {
 			fprintf (stderr, "Cannot create authenticator: %s\n",
 				 isc_result_totext (status));
