+++ dhcpctl/test.c
@@ -46,6 +46,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
+#include <string.h>
 #include <isc/result.h>
 #include "dhcpctl.h"
 
@@ -104,7 +105,7 @@ int main (argc, argv)
 	if (name) {
 		status = dhcpctl_new_authenticator (&authenticator,
 						    name, algorithm, pass,
-						    (int)strlen (pass) + 1);
+						    strlen (pass) + 1);
 		if (status != ISC_R_SUCCESS) {
 			fprintf (stderr, "Cannot create authenticator: %s\n",
 				 isc_result_totext (status));
